Pane names now reflect the connected IP address
This commit makes it so that pane names are now set to the connected IP
address. This is not always perfect as if the server is running under
some sort of proxy (in normal cases) it will only show 127.0.01.

For now this solution will work, in the future I might change it up so
it can work better with proxies somehow