Non-blocking read and write
Before this update, using the terminal was not a great experience. The
shell would accept your command, echo it back, then only print the
output after a second newline. This was because the read from stdin and
the read/write to the socket was blocking. While each was waiting, the
other couldn't update the terminal.

This is now changed for the better. Writing to the socket is
non-blocking, while keyboards are processed within the code through
poll() instead of reading one line. This makes it behave a lot more like
a real terminal.

There is one outstanding issue still. The revshell seems to echo back
the command I just typed in before giving the output. I am not yet
certain why this is, but its not a big deal. Will fix at some point